added error return code to all audio functions, added volume and panning control to channel, fixed bug in pspmappings, fixed compilerwarning in psprendersystem

+++ code/graphics/psp/psprendersystem.cpp
@@ -354,7 +354,7 @@ PSPRenderSystem::_UseLights(core::HashTable* lights)
         }
     }
 
-    for (i = lightIndex; i < MaxLights; i++)
+    for (i = lightIndex; i < (unsigned int)MaxLights; i++)
     {
         sceGuDisable(GU_LIGHT0 + i);
     }
